Move planner into a separate package
